refactor: infrastructure 내 클래스 알맞은 위치로 이동
