+++ src/main/java/com/process/clash/infrastructure/config/security/SecurityConfig.java
@@ -5,8 +5,8 @@ import com.process.clash.adapter.web.auth.service.CustomUserDetailsService;
 import com.process.clash.adapter.web.common.CommonResponse;
 import com.process.clash.adapter.web.common.ErrorResponse;
 import com.process.clash.application.common.exception.statuscode.CommonStatusCode;
-import com.process.clash.infrastructure.filter.RateLimitFilter;
-import com.process.clash.infrastructure.filter.RecaptchaFilter;
+import com.process.clash.adapter.web.filter.RateLimitFilter;
+import com.process.clash.adapter.web.filter.RecaptchaFilter;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
