--- src/main/java/com/process/clash/infrastructure/web/ServletRequestContextAdapter.java
@@ -1,35 +1,0 @@
-package com.process.clash.infrastructure.web;
-
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.stereotype.Component;
-import org.springframework.web.context.request.RequestAttributes;
-import org.springframework.web.context.request.RequestContextHolder;
-import org.springframework.web.context.request.ServletRequestAttributes;
-
-@Deprecated
-@Component
-public class ServletRequestContextAdapter {
-
-    private ServletRequestAttributes currentAttrs() {
-        RequestAttributes attrs = RequestContextHolder.getRequestAttributes();
-        if (attrs instanceof ServletRequestAttributes servletAttrs) return servletAttrs;
-        return null;
-    }
-
-    public String getIpAddress() {
-        ServletRequestAttributes attrs = currentAttrs();
-        if (attrs == null) return null;
-        HttpServletRequest req = attrs.getRequest();
-        String xff = req.getHeader("X-Forwarded-For");
-        if (xff != null && !xff.isEmpty()) {
-            return xff.split(",")[0].trim();
-        }
-        return req.getRemoteAddr();
-    }
-
-    public String getUserAgent() {
-        ServletRequestAttributes attrs = currentAttrs();
-        if (attrs == null) return null;
-        return attrs.getRequest().getHeader("User-Agent");
-    }
-}
