+++ src/test/java/com/process/clash/infrastructure/filter/RecaptchaFilterTest.java
@@ -3,6 +3,7 @@ package com.process.clash.infrastructure.filter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.process.clash.adapter.web.common.CommonResponse;
+import com.process.clash.adapter.web.filter.RecaptchaFilter;
 import com.process.clash.application.google.port.out.RecaptchaPort;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
